Refactor structure to more FSD
