+++ src/app/(pages)/generate-response/page.tsx
@@ -1,4 +1,4 @@
-import { AIChat } from "@/entities/ai-chat/ui/AIChat";
+import { AIChat } from "@/features/ai-chat";
 
 export default function GenerateResponsePage()  {
     return (
