+++ src/entities/rensponse/index.ts
@@ -1,0 +1,2 @@
+export {useResponseStore} from './model/responseStore'
+export type {AIResponse} from './model/types'
