+++ src/entities/rensponse/model/types.ts
