+++ src/entities/resume/index.ts
@@ -1,0 +1,2 @@
+export {ResumeCard} from "./ui/ResumeCard"
+export {useResume} from "./model/resumeStore"
