--- src/entities/resume/server.ts
@@ -1,1 +1,0 @@
-export {ResumeCard} from "./ui/ResumeCard"
