+++ src/features/ai-chat/index.ts
@@ -1,0 +1,1 @@
+export {AIChat} from './ui/AIChat'
