+++ src/features/response/response-list/index.ts
@@ -1,0 +1,1 @@
+export {ResponsesList} from "./ui/ResponsesList"
