+++ src/features/response/response-modal/ui/ResponseDetailsModal.tsx
@@ -5,7 +5,7 @@ import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/shared/ui/di
 import { Button } from "@/shared/ui/button";
 import { Textarea } from "@/shared/ui/textarea";
 import { Input } from "@/shared/ui/input";
-import { AIResponse } from "../model/types";
+import { AIResponse } from "../../rensponse/model/types";
 
 interface ResponseDetailsModalProps {
   open: boolean;
